chore: mockBanner 파일 경로 이동 (#304)
