--- src/mocks/mockBanner.ts
@@ -1,53 +1,0 @@
-import { Banner } from '@/types/banner'
-
-export const MOCK_BANNERS: Banner[] = [
-  {
-    id: 1,
-    imgUrl: '/images/games/game6.png',
-    title: '007 First Light',
-    genre: ['액션'],
-    tag: ['#싱글플레이', '#스파이', '#스토리'],
-    description: '제임스 본드의 새로운 미션',
-  },
-  {
-    id: 2,
-    imgUrl: '/images/games/game7.png',
-    title: '슬라임 랜처 2',
-    genre: ['액션', '어드벤처', '캐주얼', '인디'],
-    tag: ['#싱글플레이', '#멀티플레이', '#오픈월드', '#판타지'],
-    description: '괴물 사냥꾼 게롤트의 마지막 여정',
-  },
-  {
-    id: 3,
-    imgUrl: '/images/games/game8.png',
-    title: 'Clair Obscur: Expedition 33',
-    genre: ['RPG', '액션'],
-    tag: ['#싱글플레이', '#스토리', '#판타지', '#탐험'],
-    description:
-      '꿈틀거리고 출렁이는 새로운 슬라임들이 가득한 신비로운 땅, 레인보우 아일랜드로 떠나는 모험을 계속하세요.',
-  },
-  {
-    id: 4,
-    imgUrl: '/images/games/game9.png',
-    title: 'The Legend of Zelda: Tears of the Kingdom',
-    genre: ['액션', '대규모 멀티플레이어'],
-    tag: ['#싱글플레이', '#오픈월드', '#독점'],
-    description:
-      '닌텐도 스위치™용 <젤다의 전설: 왕국의 눈물>에서 하이룰의 땅과 하늘을 가로지르는 장대한 모험이 여러분을 기다립니다. ',
-  },
-  {
-    id: 5,
-    imgUrl: '/images/games/game10.png',
-    title: 'DiRT Rally',
-    genre: ['시뮬레이션', '스포츠', '레이싱'],
-    tag: [
-      '#싱글플레이',
-      '#멀티플레이',
-      '#고난이도',
-      '#무료플레이',
-      '#현실적인',
-    ],
-    description:
-      '아슬아슬한 속도로 질주하는 위험천만한 도로 위에서 손에 땀을 쥐게 하는 레이싱의 짜릿함을 완벽하게 재현합니다.',
-  },
-]
